refactor(MITE-1): Modularização das Views
