+++ .env
@@ -1,0 +1,5 @@
+DB_NAME=Teste
+DB_USER=root
+DB_PASSWORD=12345
+DB_HOST=localhost
+DB_PORT=3306
