+++ budibase_aux_screens/budibase_aux_screens/wsgi.py
@@ -1,0 +1,16 @@
+"""
+WSGI config for budibase_aux_screens project.
+
+It exposes the WSGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/5.0/howto/deployment/wsgi/
+"""
+
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'budibase_aux_screens.settings')
+
+application = get_wsgi_application()
