+++ budibase_aux_screens/screens/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ScreensConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'screens'
