--- infra/budibase_aux_screens/screens/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('<int:c_id>', views.main, name=''),
-]
